clean: Rename the component files.
