+++ src/app/layout.tsx
@@ -1,6 +1,6 @@
 import './globals.css';
-import Header from '@/components/header';
-import Footer from '@/components/footer';
+import Header from '@/components/Header';
+import Footer from '@/components/Footer';
 
 export const metadata = {
   title: '株式会社前野企画',
