+++ src/app/page.tsx
@@ -1,4 +1,4 @@
-import Hero from '@/components/hero';
+import Hero from '@/components/Hero';
 
 export default function Home() {
   return (
