+++ src/components/Footer.tsx
@@ -1,4 +1,4 @@
-import Logo from '@/components/logo';
+import Logo from '@/components/Logo';
 
 export default function Footer() {
   return (
